Move R3 replay mask into the router primitive
